General Code
This updated version of LEiDA also calculates the lifetimes and probabilities of each state.
In addition, it now ensures that the leading eigenvectors returned by the function [V, ~]=eigs(iFC,1) is always the most negative. Otherwise Matlab randomly returns V or -V, which alters the results.
Also, we don't use the dunn's score anymore to estimate the best number of states, but instead search for the number of states that better distinguish between conditions (i.e. rest vs task, or disease vs control)
We include an example BOLD dataset.